pyproject.toml
[project]
name = "rag-pipline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "llama-index>=0.12.48",
    "llama-index-embeddings-openai>=0.3.1",
    "llama-index-llms-groq>=0.3.2",
    "llama-index-vector-stores-qdrant>=0.6.1",
    "pymupdf4llm>=0.0.26",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.14.3",
    "streamlit>=1.46.1",
]
